+++ src/components/Button/index.ts
@@ -1,3 +1,3 @@
 export { Button as default } from './Button';
 export type { ButtonProps } from './Button.types';
-export type { ButtonColor, ButtonColorProp, ButtonSize, ButtonVariant } from './Button.types';
+export type { ButtonColor, ButtonColorProp, ButtonSize, ButtonSizeProp, ButtonVariant } from './Button.types';